chapter

+++ src/main/java/com/insat/maktabti/DAO/ChapterDao.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface ChapterDao extends JpaRepository<Chapter,Integer> {
-    Chapter findById(int id);
+    Chapter findById(long id);
 }
